+++ public/index.html
@@ -32,6 +32,6 @@
     <!-- Количество фильмов -->
   </section>
 </footer>
-<script src="../src/main.js"></script>
+<script type="module" src="./bundle.js"></script>
 </body>
 </html>
