+++ src/components/films-extra.js
@@ -1,0 +1,9 @@
+const createFilmsListExtraTemplate = function (title) {
+  return (
+    `<section class="films-list--extra">
+      <h2 class="films-list__title">${title}</h2>
+      <div class="films-list__container"></div>
+    </section>`);
+};
+
+export default createFilmsListExtraTemplate;
