+++ src/components/footer-statistic.js
@@ -1,0 +1,9 @@
+const createFooterStatisticTemplate = function () {
+  return (
+    `<section class="footer__statistics">
+        <p>130 291 movies inside</p>
+      </section>`
+  );
+};
+
+export default createFooterStatisticTemplate;
