+++ src/components/load-button.js
@@ -1,0 +1,7 @@
+const createLoadButtonTemplate = function () {
+  return (
+    `<button class="films-list__show-more">Show more</button>`
+  );
+};
+
+export default createLoadButtonTemplate;
